backend/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
textblob==0.17.1
elevenlabs==0.2.24
bcrypt==4.0.1
PyJWT==2.8.0
gunicorn==21.2.0
python-dateutil==2.8.2
requests==2.31.0
cryptography==41.0.4
pytest==7.4.3
black==23.11.0
flake8==6.1.0
supabase==2.0.3 